size

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int lli;
 	float fl;
 
-	printf("size of a char: %lu byte(s)\n", sizeof(ch));
-	printf("size of an int: %lu byte(s)\n", sizeof(in));
-	printf("size of long int: %lu byte(s)\n", sizeof(li));
-	printf("size of long long int: %lu byte(s)\n", sizeof(lli));
-	printf("size of float: %lu byte(s)\n", sizeof(fl));
+	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(ch));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(in));
+	printf("size of long int: %lu byte(s)\n", (unsigned long)sizeof(li));
+	printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(lli));
+	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(fl));
 	return (0);
 }
